requirements.txt
asgiref==3.4.1
Django==3.2.6
pytz==2021.1
sqlparse==0.4.1
dj-database-url==0.5.0
psycopg2-binary==2.9.1
dj-database-url
drf_yasg
djangorestframework
django-rest-authtoken
djangorestframework-jsonapi
django-location-field
django_heroku